Add FsProber class
While inotify provides informations about the changes occuring in the inode it
watches, such watch is not propagated to its subdirectories. For example, if the
command `mkdir -p somedir/anotherdir/yetanotherdir` is issued in a watched
directory, inotify will only report about somedir's creation

FsProber provides methods to probe the filesystem and obtain informations about
the creations and deletions that occurred, and will be used to propagate
inotify's watch in the lower layers of the filesystem.